Fix an edge case of tie determination
There was a bug where if the combination of squares in a grid consisted
of all 3 values (X, O, & ties) that it would not be able to determine
if it were a tie. This change requires that the whole grid be filled
before it can determine a tie, but it'll catch those situations it did
not before.